requirements.txt
Django>=3.2
gunicorn>=20.1.0  # For production WSGI server
monero>=0.6.0  # Monero RPC clientdjango-jazzmin
python-dotenv
python-bitcoinrpc
requests